+++ pkg/resource/events-logger-config/alloy-events-config_test.go
@@ -69,7 +69,7 @@ func TestGenerateAlloyEventsConfig(t *testing.T) {
 				},
 			}
 
-			config, err := generateAlloyEventsConfig(cluster, tc.includeNamespaces, tc.excludeNamespaces, tc.installationName, false, tc.tracingEnabled, "<tempo-url>")
+			config, err := generateAlloyEventsConfig(cluster, tc.includeNamespaces, tc.excludeNamespaces, tc.installationName, false, tc.tracingEnabled, "<tempo-url>", []string{"giantswarm"})
 			if err != nil {
 				t.Fatalf("Failed to generate alloy config: %v", err)
 			}