+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `otlp` receiver and exporters to alloy-events.
+- Add multi-tenancy for traces in alloy-events.
 
 ## [0.30.0] - 2025-09-03